+++ src/VortexStepMethod.jl
@@ -12,13 +12,12 @@ using LaTeXStrings
 using NonlinearSolve
 using Interpolations: linear_interpolation, Line, Extrapolation
 using Serialization
-using Distributed
 using SharedArrays
 using BenchmarkTools
 
 # Export public interface
 export Wing, Section, KiteWing
-export WingAerodynamics
+export BodyAerodynamics
 export Solver, solve
 export calculate_results, solve_circulation_distribution
 export add_section!, set_va!
@@ -39,7 +38,15 @@ const MVec3    = MVector{3, Float64}
 Position vector, either a `MVec3` or a `Vector` for use in function signatures.
 """
 const PosVector=Union{MVec3, Vector, SizedVector{3, Float64, Vector{Float64}}}
+
+"""
+   const VelVector=Union{MVec3, Vector}
+
+Velocity vector, either a `MVec3` or a `Vector` for use in function signatures.
+"""
 const VelVector=Union{MVec3, Vector, SizedVector{3, Float64, Vector{Float64}}}
+
+@enum Model VSM LLT
 
 abstract type AbstractWing end
 
@@ -53,7 +60,7 @@ include("kite_geometry.jl")
 include("filament.jl")
 include("panel.jl")
 include("wake.jl")
-include("wing_aerodynamics.jl")
+include("body_aerodynamics.jl")
 include("solver.jl")
 
 # include plotting